refactor: move storage location to separate module
